Algorithmia website commit

+++ src/crossword_extractor.py
@@ -169,8 +169,8 @@ def get_grid_count(input):
 def is_black_square(input, grid_count, row, col):
     sp = float(input.shape[0]) / float(grid_count);
     # get actual row/col pixel
-    r = int(double(row) * sp + sp / 2);
-    c = int(double(col) * sp + sp / 2);
+    r = int(float(row) * sp + sp / 2);
+    c = int(float(col) * sp + sp / 2);
 
     (_, tmp) = cv2.threshold(input, 128., 255., cv2.THRESH_BINARY);
     dim = int(sp/4)
